backend/organizations/requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.4
djangorestframework==3.15.1
jsonschema==4.17.3
pyrsistent==0.20.0
sqlparse==0.5.0
